Split TdsParserHelperClasses into separate files Consolidate partial classes Add references to Parser namespace
