+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/AlwaysEncryptedHelperClasses.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -191,7 +191,7 @@ namespace Microsoft.Data.SqlClient
         }
     }
 
-    /// <summary> 
+    /// <summary>
     /// Represents a table with various CEKs used in a resultset. Each entry corresponds to one (unique) CEK. The CEK
     /// may have been encrypted using multiple master keys (giving us multiple CEK values). All these values form one single
     /// entry in this table.
@@ -226,17 +226,6 @@ namespace Microsoft.Data.SqlClient
             {
                 return keyList.Length;
             }
-        }
-    }
-
-    internal sealed partial class _SqlMetaDataSet
-    {
-        internal readonly SqlTceCipherInfoTable cekTable; // table of "column encryption keys" used for this metadataset
-
-        internal _SqlMetaDataSet(int count, SqlTceCipherInfoTable cipherTable)
-            : this(count)
-        {
-            cekTable = cipherTable;
         }
     }
 
@@ -425,44 +414,6 @@ namespace Microsoft.Data.SqlClient
         internal bool IsAlgorithmInitialized()
         {
             return _sqlClientEncryptionAlgorithm != null ? true : false;
-        }
-    }
-
-    internal partial class SqlMetaDataPriv
-    {
-        internal bool isEncrypted; // TCE encrypted?
-        internal SqlMetaDataPriv baseTI;   // for encrypted columns, represents the TYPE_INFO for plaintext value
-        internal SqlCipherMetadata cipherMD; // Cipher related metadata for encrypted columns.
-
-        /// <summary>
-        /// Is the algorithm handle for the cipher encryption initialized ?
-        /// </summary>
-        /// <returns></returns>
-        internal bool IsAlgorithmInitialized()
-        {
-            if (cipherMD != null)
-            {
-                return cipherMD.IsAlgorithmInitialized();
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        /// Returns the normalization rule version byte.
-        /// </summary>
-        /// <returns></returns>
-        internal byte NormalizationRuleVersion
-        {
-            get
-            {
-                if (cipherMD != null)
-                {
-                    return cipherMD.NormalizationRuleVersion;
-                }
-
-                return 0x00;
-            }
         }
     }
