+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ManagedSni/SniLoadHandle.netcore.cs
@@ -5,6 +5,7 @@
 #if NET
 
 using System;
+using Microsoft.Data.SqlClient.Parser;
 
 #nullable enable
