+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/PreLoginHandshakeStatus.cs
@@ -1,0 +1,13 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable enable
+
+namespace Microsoft.Data.SqlClient.Parser;
+
+internal enum PreLoginHandshakeStatus
+{
+    Successful,
+    InstanceFailure
+}
