+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/SqlMetaDataXmlSchemaCollection.cs
@@ -1,0 +1,22 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.Data.SqlClient.Parser;
+
+internal sealed class SqlMetaDataXmlSchemaCollection
+{
+    internal string Database;
+    internal string OwningSchema;
+    internal string Name;
+
+    public void CopyFrom(SqlMetaDataXmlSchemaCollection original)
+    {
+        if (original != null)
+        {
+            Database = original.Database;
+            OwningSchema = original.OwningSchema;
+            Name = original.Name;
+        }
+    }
+}
