+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/SqlReturnValue.cs
@@ -1,0 +1,21 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.Data.SqlClient.Parser;
+
+internal sealed class SqlReturnValue : SqlMetaDataPriv
+{
+    // @TODO: Make auto properties (and rename to match)
+    internal string parameter;
+    internal readonly SqlBuffer value;
+
+    #if NETFRAMEWORK
+    internal ushort parmIndex;      //2005 or later only
+    #endif
+
+    internal SqlReturnValue() : base()
+    {
+        value = new SqlBuffer();
+    }
+}
